cdk/mwaa_authx_lambda_code/requirements.txt
pyjwt
requests
urllib3
cryptography
cffi